# Change "destroyed" segments / synapses into an implementation detail