… src/nupic/experimental/ExtendedTemporalMemory.cpp
… src/test/unit/algorithms/ConnectionsTest.cpp
… src/test/unit/algorithms/TemporalMemoryTest.cpp
… src/test/unit/experimental/ExtendedTemporalMemoryTest.cpp