Fixes #1064 

Here's the matching Python change: https://github.com/numenta/nupic/pull/3322. Refer to that comment to see how this changes the TM output.

## Changes to data storage in the Connections

Previously, the `Connections` stored its data in lots of small vectors. It had:

- A vector of `CellData` structs
  - Each holding a vector of `SegmentData` structs, some of which might be destroyed
    - Each holding a vector of `SynapseData` structs, some of which might be destroyed

The path to a segment was a `(cellIdx, segmentIdx)` tuple, and the path to a synapse was a `(cellIdx, segmentIdx, synapseIdx)` tuple. Destroyed synapses lived on their segment and couldn't be recycled by other segments. Same with destroyed segments and their cell.

Now the `Connections` stores this data in 3 flat lists:

- A vector of `CellData` structs
- A vector of `SegmentData` structs, some of which might be destroyed
- A vector of `SynapseData` structs, some of which might be destroyed

The path to a segment or to a synapse is simply its index in this list. Destroyed synapses can now be recycled by any segment, so the total number of allocated `SynapseData` structs is smaller. Same with destroyed segments and their cell.

The CellData no longer stores SegmentData, and the SegmentData no longer stores SynapseData. Instead, these just keep lists of paths to their segments / synapses. These lists are very convenient -- they are the correct return values for `connections.segmentsForCell` and `connections.synapsesForSegment`. Now these methods can simply return a reference to the existing list, so they cost nothing.

The `Connections` flat lists of SegmentData and SynapseData are essentially memory heaps. We essentially have implemented a segment heap and a synapse heap, and a `Segment` and a `Synapse` are essentially pointers into these heaps. The main reason we maintain these heaps isn't for perf -- it's because we don't want to invalidate `Segment` and `Synapse` paths.

I haven't done any work to optimize these heaps, but in the future we might want a `connections.defrag()` method which sorts the data in the heap, grouping all synapses on a segment into a contiguous group, which will make much better use of CPU cache. We wouldn't want to do this automagically, since it would invalidate any existing `Segment` and `Synapse` "pointers", or it would require adding another layer of indirection.

## Perf

**Before**

~~~
> python scripts/temporal_memory_performance_benchmark.py -t hotgym 20_hotgym

Test: hotgym

tm_cpp: 0.707339s


Test: hotgym, 20 times

tm_cpp: 41.275666s
~~~

**After**

~~~
> python scripts/temporal_memory_performance_benchmark.py -t hotgym 20_hotgym

Test: hotgym

tm_cpp: 0.555045s


Test: hotgym, 20 times

tm_cpp: 32.429480s
~~~

So it's ~27% faster.

It's faster because this change allows us to make `segmentsForCell` and `synapsesForSegment` instant. They just return a reference to an already-existing vector.