@mrcslws 2016-09-19
This is ready for review. @scottpurdy